+++ src/decorator.tsx
@@ -1,8 +1,16 @@
 import React, {Component} from "react";
 import {THEME_STATE, ThemeState} from "./reducer";
-import {SET_STICKER_TYPE, SET_THEME, STICKER_UPDATED, TOGGLE_FONT, TOGGLE_STICKER, TOGGLE_WALLPAPER,} from "./settings";
+import {
+  SET_STICKER_TYPE,
+  SET_THEME,
+  STICKER_UPDATED,
+  SYSTEM_THEME_CHANGED,
+  TOGGLE_FONT,
+  TOGGLE_STICKER,
+  TOGGLE_WALLPAPER,
+} from "./settings";
 import path from "path";
-import {resolveLocalStickerPath} from "./StickerUpdateService";
+import {attemptToUpdateSticker, resolveLocalStickerPath} from "./StickerUpdateService";
 import {ipcRenderer} from "electron";
 
 const passProps = (uid: any, parentProps: any, props: any) =>
@@ -67,6 +75,29 @@ const createCacheBuster = () => new Date().valueOf().toString(32);
 
 let initialized = false;
 
+const mediaQuery = "(prefers-color-scheme: dark)";
+
+/**
+ * Just in case you forget, all the terminal ui changes
+ * happen in the Redux Store that only lives inside the
+ * Electron Window app.
+ *
+ * So things outside the decorator don't have access to the Redux Store.
+ * Which is a pain in the ass. So stuff outside of Redux uses config files.
+ *
+ * Things like the settings emit events that this decorator listens to and
+ * emits the corresponding events to create a new state.
+ *
+ * Anytime that those changes happen, I also need to reload the configuration
+ * to force update all the other crap in index.ts, which causes things to
+ * change.
+ *
+ * This code base looks stupid, because it is stupid.
+ *
+ * Love,
+ *
+ * Past Alex
+ */
 export const decorateTerm = (Term: any) => {
   let cacheBuster: string = createCacheBuster();
   return class TerminalDecorator extends Component<any, StickerState> {
@@ -86,6 +117,25 @@ export const decorateTerm = (Term: any) => {
       ).replace(path.sep, "/");
       return `${localStickerPath}?time=${cacheBuster}`;
     }
+
+    private mediaChangeListener = TerminalDecorator.handleMediaChange.bind(this);
+
+    private static handleMediaChange() {
+      window.store.dispatch({
+        type: SYSTEM_THEME_CHANGED,
+        payload: {
+          isDark: TerminalDecorator.isDark(),
+        }
+      });
+      // this refreshes the entire application so
+      // the new theme gets picked up.
+      window.store.dispatch(reloadConfig(window.config.getConfig()));
+    }
+
+    public static isDark() {
+      return window.matchMedia(mediaQuery).matches;
+    }
+
 
     componentDidMount() {
       if (!initialized) {
@@ -112,6 +162,11 @@ export const decorateTerm = (Term: any) => {
             type: TOGGLE_WALLPAPER,
           });
         });
+        // dispatch to initialize system state.
+        TerminalDecorator.handleMediaChange();
+        window
+            .matchMedia(mediaQuery)
+            .addEventListener("change", this.mediaChangeListener);
         initialized = true;
       }
     }
@@ -136,6 +191,22 @@ export const decorateTerm = (Term: any) => {
       ) {
         this.setState({ imageLoaded: false });
         cacheBuster = createCacheBuster();
+      }
+
+      if(themeState.activeTheme.information.id !=
+        nextThemeState.activeTheme.information.id) {
+        // if user updates config, that causes a potential theme change
+        // eg. if system is dark and they change dark theme. The config
+        // change causes the active theme to change. When the theme
+        // changes on config reload, the CSS has already been decorated.
+        // CSS decoration is how the terminal is themed. So the terminal
+        // was decorated with the previous theme (the initial config reload
+        // triggered by the user). So dispatch a config reload again,
+        // to pick up the latest theme changes.
+        window.store.dispatch(reloadConfig(window.config.getConfig()));
+        attemptToUpdateSticker().then(() => {
+          window.store.dispatch(reloadConfig(window.config.getConfig()))
+        });
       }
     }